requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
python-dotenv
sentence-transformers
torch
chromadb
groq
